requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
pathlib>=1.0.0
requests>=2.0.0
